改角度之后的第一次成功version

+++ HARDWARE/motor/motor.c
@@ -70,8 +70,8 @@ void set_angle(float x_angle, float y_angle)
 	
 	
 	
-	PWM_x = x_angle / 180 * 2000 + 500;
-	PWM_y = y_angle / 180 * 2000 + 500;
+	PWM_x = x_angle / (float)180 * 2000 + 500;
+	PWM_y = y_angle / (float)180 * 2000 + 500;
 	
 	TIM_SetCompare1(TIM4,PWM_x);	//修改比较值，修改占空比      输出pid计算值
 	TIM_SetCompare2(TIM4,PWM_y);	//修改比较值，修改占空比
